Added HLF support for ProductService

+++ pumpk1n_backend/Services/Products/ProductService.cs
@@ -4,9 +4,11 @@ using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
 using Microsoft.EntityFrameworkCore;
+using pumpk1n_backend.Exceptions.Chain;
 using pumpk1n_backend.Exceptions.Others;
 using pumpk1n_backend.Exceptions.Products;
 using pumpk1n_backend.Models;
+using pumpk1n_backend.Models.ChainTransferModels.Products;
 using pumpk1n_backend.Models.DatabaseContexts;
 using pumpk1n_backend.Models.Entities.Products;
 using pumpk1n_backend.Models.ReturnModels.Products;
@@ -18,14 +20,16 @@ namespace pumpk1n_backend.Services.Products
     {
         private readonly DatabaseContext _context;
         private readonly IMapper _mapper;
+        private readonly IProductChainService _productChainService;
 
         private const string DefaultImageUrl =
             "https://ssl-product-images.www8-hp.com/digmedialib/prodimg/lowres/c05962484.png";
 
-        public ProductService(DatabaseContext context, IMapper mapper)
+        public ProductService(DatabaseContext context, IMapper mapper, IProductChainService productChainService)
         {
             _context = context;
             _mapper = mapper;
+            _productChainService = productChainService;
         }
 
         public async Task<ProductReturnModel> AddProduct(ProductInsertModel model)
@@ -39,6 +43,15 @@ namespace pumpk1n_backend.Services.Products
                     _context.Products.Add(product);
 
                     await _context.SaveChangesAsync();
+                    
+                    var chainModel = new ChainProductTransferModel
+                    {
+                        Id = product.Id.ToString(),
+                        CreatedDate = product.AddedDate.ToBinary().ToString(),
+                        Hash = product.ComputeHash().ToString()
+                    };
+                    await _productChainService.AddProduct(chainModel);
+                    
                     transaction.Commit();
 
                     return _mapper.Map<Product, ProductReturnModel>(product);
@@ -62,10 +75,25 @@ namespace pumpk1n_backend.Services.Products
                     if (product == null)
                         throw new ProductNotFoundException();
                     
+                    var productChainInfo = await _productChainService.GetProduct(product.Id);
+                    if (productChainInfo == null)
+                        throw new DataNotFoundInChainException();
+                    if (long.Parse(productChainInfo.Hash) != product.ComputeHash())
+                        throw new ChainCodeDataNotInSyncException();
+                    
                     _mapper.Map(model, product);
 
                     _context.Products.Update(product);
                     await _context.SaveChangesAsync();
+                    
+                    var chainModel = new ChainProductTransferModel
+                    {
+                        Id = product.Id.ToString(),
+                        CreatedDate = product.AddedDate.ToBinary().ToString(),
+                        Hash = product.ComputeHash().ToString()
+                    };
+                    await _productChainService.AddProduct(chainModel);
+                    
                     transaction.Commit();
 
                     return _mapper.Map<Product, ProductReturnModel>(product);
@@ -88,9 +116,17 @@ namespace pumpk1n_backend.Services.Products
 
                     if (product == null)
                         throw new ProductNotFoundException();
+                    
+                    var productChainInfo = await _productChainService.GetProduct(product.Id);
+                    if (productChainInfo == null)
+                        throw new DataNotFoundInChainException();
+                    if (long.Parse(productChainInfo.Hash) != product.ComputeHash())
+                        throw new ChainCodeDataNotInSyncException();
 
                     _context.Products.Remove(product);
                     await _context.SaveChangesAsync();
+                    
+                    await _productChainService.DeleteProduct(productId);
                     transaction.Commit();
                 }
                 catch (Exception)
@@ -107,6 +143,12 @@ namespace pumpk1n_backend.Services.Products
 
             if (product == null)
                 throw new ProductNotFoundException();
+            
+            var productChainInfo = await _productChainService.GetProduct(product.Id);
+            if (productChainInfo == null)
+                throw new DataNotFoundInChainException();
+            if (long.Parse(productChainInfo.Hash) != product.ComputeHash())
+                throw new ChainCodeDataNotInSyncException();
 
             if (string.IsNullOrEmpty(product.Image) || string.IsNullOrWhiteSpace(product.Image))
                 product.Image = DefaultImageUrl;
@@ -127,10 +169,18 @@ namespace pumpk1n_backend.Services.Products
                 .OrderByDescending(p => p.AddedDate)
                 .Skip(startAt)
                 .Take(count).ToListAsync();
-            
+
             foreach (var product in products)
+            {
+                var productChainInfo = await _productChainService.GetProduct(product.Id);
+                if (productChainInfo == null)
+                    throw new DataNotFoundInChainException();
+                if (long.Parse(productChainInfo.Hash) != product.ComputeHash())
+                    throw new ChainCodeDataNotInSyncException();
+                
                 if (string.IsNullOrEmpty(product.Image) || string.IsNullOrWhiteSpace(product.Image))
                     product.Image = DefaultImageUrl;
+            }
 
             var totalCount = await _context.Products
                 .Where(p => p.Name.Contains(filterModel.Name, StringComparison.InvariantCultureIgnoreCase) &&
@@ -156,11 +206,27 @@ namespace pumpk1n_backend.Services.Products
                     var product = await _context.Products.FirstOrDefaultAsync(p => p.Id == productId);
                     if (product == null)
                         throw new ProductNotFoundException();
+                    
+                    var productChainInfo = await _productChainService.GetProduct(product.Id);
+                    if (productChainInfo == null)
+                        throw new DataNotFoundInChainException();
+                    if (long.Parse(productChainInfo.Hash) != product.ComputeHash())
+                        throw new ChainCodeDataNotInSyncException();
 
                     product.Deprecated = status;
 
                     _context.Products.Update(product);
                     await _context.SaveChangesAsync();
+                    
+                    var chainModel = new ChainProductTransferModel
+                    {
+                        Id = product.Id.ToString(),
+                        CreatedDate = product.AddedDate.ToBinary().ToString(),
+                        Hash = product.ComputeHash().ToString()
+                    };
+                    await _productChainService.AddProduct(chainModel);
+                    
+                    transaction.Commit();
                 }
                 catch (Exception)
                 {
@@ -179,11 +245,26 @@ namespace pumpk1n_backend.Services.Products
                     var product = await _context.Products.FirstOrDefaultAsync(p => p.Id == productId);
                     if (product == null)
                         throw new ProductNotFoundException();
+                    
+                    var productChainInfo = await _productChainService.GetProduct(product.Id);
+                    if (productChainInfo == null)
+                        throw new DataNotFoundInChainException();
+                    if (long.Parse(productChainInfo.Hash) != product.ComputeHash())
+                        throw new ChainCodeDataNotInSyncException();
 
                     product.OutOfStock = !status;
 
                     _context.Products.Update(product);
                     await _context.SaveChangesAsync();
+                    
+                    var chainModel = new ChainProductTransferModel
+                    {
+                        Id = product.Id.ToString(),
+                        CreatedDate = product.AddedDate.ToBinary().ToString(),
+                        Hash = product.ComputeHash().ToString()
+                    };
+                    await _productChainService.AddProduct(chainModel);
+                    
                     transaction.Commit();
                 }
                 catch (Exception)
